refactor: SocketInformation 위치 수정
