+++ src/conversation-events/conversation-events.module.ts
@@ -6,7 +6,6 @@ import { RoomsModule } from 'src/rooms/rooms.module';
 import { ConversationEventsService } from './conversation-events.service';
 import { RoomEventModule } from 'src/rooms/item/room-event/room-event.module';
 import { RoomSocketModule } from './room-socket/room-socket.module';
-import { SocketInformationModule } from './socket-information/socket-information.module';
 
 @Module({
   imports: [
@@ -15,7 +14,6 @@ import { SocketInformationModule } from './socket-information/socket-information
     RoomsModule,
     RoomEventModule,
     RoomSocketModule,
-    SocketInformationModule,
   ],
   providers: [ConversationEventsGateway, ConversationEventsService],
 })
