--- src/conversation-events/socket-information/socket-information.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { SocketInformationService } from './socket-information.service';
-
-@Module({
-  providers: [SocketInformationService],
-  exports: [SocketInformationService],
-})
-export class SocketInformationModule {}
