--- src/conversation-events/socket-information/socket-information.service.ts
@@ -1,4 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class SocketInformationService {}
