+++ src/rooms/item/room-event/socket-information.ts
@@ -1,5 +1,8 @@
 import { logger } from 'src/rooms/item/room-event';
-import { RoomSocket, SocketStatus } from '../room-socket';
+import {
+  RoomSocket,
+  SocketStatus,
+} from '../../../conversation-events/room-socket';
 
 //TODO: 여기 다른 것 추가해야하는지 front에게 묻기
 export class SocketInformation {
